+++ packages/ui-core/src/components/core/index.ts
@@ -1,9 +1,13 @@
 import ErrorInfo from './ErrorInfo';
 import Loading from './Loading';
 import Page from './Page';
+import TableHeader from './TableHeader';
+import TableToolbar from './TableToolbar';
 
 export {
     ErrorInfo,
     Loading,
     Page,
+    TableHeader,
+    TableToolbar
 };
