--- packages/ui-core/src/components/data-access/UsersTable/shared.ts
@@ -1,7 +1,0 @@
-export const rows = [
-  { id: 'firstname', label: 'First Name' },
-  { id: 'lastname', label: 'Last Name' },
-  { id: 'username', label: 'Username' },
-  { id: 'role', label: 'Role' },
-  { id: 'created', label: 'Created' },
-];
