+++ packages/ui-core/src/components/framework/AppWrapper/Footer.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import Typography from '@material-ui/core/Typography';
+
+const Footer: React.FC = () => {
+    return (
+        <Typography
+            component="footer"
+            align="center"
+        >
+            Copyright &copy; 2019
+        </Typography>
+    );
+};
+
+export default Footer;
